Refactor firebase init out of login hook.
