+++ hooks/useAnonymousLoginIfNeeded.ts
@@ -1,28 +1,16 @@
-import { getApps, initializeApp } from 'firebase/app';
+import { getApps } from 'firebase/app';
 import { Auth, getAuth, onAuthStateChanged, signInAnonymously, User } from 'firebase/auth';
 import { useCallback, useEffect, useState } from 'react';
 import { Nullable } from '../types/utils';
 
 /**
- * Hook that handles anonymous log in / register if not already logged in. This will also handle firebase
- * initialization if needed.
+ * Hook that handles anonymous log in / register if not already logged in.
  *
  * Note that this hook should not be needed in most cases as it is already called in `_app.tsx`.
  * Call this hook when firebase auth functions are required (e.g. `getAuth`, `signOut`, `signInWith`...)
  * Do not call this hook to access user object (Just use `AuthContext.Consumer`)
  */
 export default function useAnonymousLoginIfNeeded() {
-  const initFirebase = useCallback(() => {
-    initializeApp({
-      apiKey: process.env.NEXT_PUBLIC_FIREBASE_API_KEY,
-      authDomain: process.env.NEXT_PUBLIC_FIREBASE_AUTH_DOMAIN,
-      projectId: process.env.NEXT_PUBLIC_FIREBASE_PROJECT_ID,
-      storageBucket: process.env.NEXT_PUBLIC_FIREBASE_STORAGE_BUCKET,
-      messagingSenderId: process.env.NEXT_PUBLIC_FIREBASE_MESSAGING_SENDER_ID,
-      appId: process.env.NEXT_PUBLIC_FIREBASE_APP_ID,
-    });
-  }, [initializeApp]);
-
   const [currentUser, setCurrentUser] = useState<User>();
   const [auth, setAuth] = useState<Auth>();
 
@@ -40,7 +28,7 @@ export default function useAnonymousLoginIfNeeded() {
 
   useEffect(() => {
     if (!getApps().length) {
-      initFirebase();
+      throw new Error('Firebase app has not been initiated.');
     }
 
     const auth = getAuth();
